requirements.txt
tweepy
python-dotenv
requests
beautifulsoup4
ollama